+++ CHANGELOG.md
@@ -12,7 +12,7 @@ v0.15.15 (02.11.2015)
 v0.15.14 (21.10.2015)
  * SimpleMarker has a new method **prepareTemplate()** as an entry point for child classes
  * tx_rnbase_view_List integrates markers provided by controller class
- * tx_rnbase_action_BaseIOC has new utility method createURI()
+ * tx_rnbase_action_BaseIOC has new utility method createLink()
 
 v0.15.13 (14.10.2015)
  * add missing return statement in language utility